Consume isn't tracked.
Don't count calls to consume on a stream as a read call, thereby
dirtying the re-execution chain.  This is problematic as we're not sure
what the semantics on reading a stream of lattices is.

For example, given a stream of G-Sets.  If the first G-set is consumed,
and a LWW-register is updated as a result, when we consume the second
G-set, do we re-evaluate the entire stream again, as the second write to
the LWW-register causes the original write to be marked for
re-execution.

For now, only explicit reads of values are tracked.