Introduce notify function.
Notify is responsible for storing the scope for a function, and
re-executing the function when the scope is modified.  For example,
when computing the Cartesian product, you want to re-execute whenever
either of the sources advances monotonically.  Notify is responsible
for tracking the changes to each variable and propagating the state
change forward.