Alter program execution model.
Alter the program execution model and lasp_program behavior to take a
state object, which can contain the magic binary/reference object
requried for writing to backends which are implemented as NIFs.

This reduces the usefulness of the parse transformation, because code
must explicitly be written using the store parameter, but fixing the
transformation is a known piece of future work.