Extract out lasp_process.
Properly supervise all Lasp processes through a new abstraction called
the Lasp "process."  A Lasp "process" is a process that never
terminates; it reads input values from one or more CRDTs, transforms
them, and creates output values, both streams, of which are fully
monotonic.  This is the core abstraction in Lasp, and represents a
single node in the dataflow execution graph.