Mutations are done with closures.
Perform mutations using closures; the ensures that there are no race conditions where we read, the value changes, and we write back.  Combined with a storage_backend that implements the gen_server behavior, this prevents race conditions observed where read processes block waiting for an input, but write processes erase that they are waiting because they write from an earlier read observation.