Add lasp_process:start_dag_link/3
Additional support for lasp processes in 3 steps: Read, Transform and
Write.

Now, instead of passing a single function that computes the changes and
then writes them to the appropiate CRDT, we pass two functions, one that
computes the new value, and another that writes the value.

Todo: Modify the lasp_process state