Monitor lasp processes
Since lasp processes might terminate (as they are marked transient) or
get restarted, we have to be able to update the labels in the edges that
represent those processes.

The lasp_dependence_dag module monitors all pids that get added as
labels in an edge of the graph. When one of those monitored processes
dies, for whatever reason, we remove all the edges that reference that
pid.

Those processes that get restarted will join the dag again as soon as
they execute their init function.