Repair garbage collection routine.
Fix several problems with the garbage collection mechanism:

1.) Garbage collection would never be invoked because the garbage
collection counter would never advance as a result of the periodic timer
firing.

2.) If all members of the acknowledgement map were pruned, the object
would never be written back to the data store with the empty map, it
would ignore the pruning procedure.

3.) Reading and writing the results were seperate: objects were pruned
and later written asynchronously: this leads to a race condition where
the map is altered and the change lost because of non-atomic writes.

4.) Asynchronous read/writes were also interleaved with new updates,
causing the map to be non-atomically updated leading to dropped read and
write operations.