Fix failure to converge in reactive mode.
When a node creates a CRDT by declaring it, but without issuing an
actual mutation, it will create a bottom delta that *has* to be shipped
to the clients for them to bootstrap the object.

This is particularly problematic when the following happens:

1.) Server creates the CRDTs (at bottom.)
2.) Server sends them to client.
3.) Client performs the first mutation.

In this example, if the empty delta never gets sent to the clients, the
clients will never update the CRDT, therefore, they won't synchronize
the objects with the server, which will prevent the test from ever
converging.