Prevent non-existent vertex crash.
If a user tries to read an object that hasn't been declared yet, there will not be a vertex for the object.  Instead of eagerly creating vertexes for every object read, instead, just ignore the error and leave the graph unchanged and return a bottom value for the requested type.